pyproject.toml
[project]
name = "ai-web-explorer"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Milos Svana", email = "milos.svana@gmail.com" }
]
dependencies = [
    "playwright>=1.42.0",
    "openai>=1.14.2",
    "pyyaml>=6.0.1",
    "beautifulsoup4>=4.12.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
explore = "ai_web_explorer:main"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=24.3.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ai_web_explorer"]
